+++ src/Sulu/Bundle/ContactBundle/Controller/ContactController.php
@@ -12,11 +12,13 @@ namespace Sulu\Bundle\ContactBundle\Controller;
 
 use DateTime;
 use FOS\RestBundle\Controller\Annotations\Get;
+use Hateoas\Configuration\Exclusion;
 use JMS\Serializer\SerializationContext;
 use Sluggable\Fixture\Position;
 use Sulu\Bundle\ContactBundle\Contact\AbstractContactManager;
 use Sulu\Bundle\ContactBundle\Entity\Account;
 use Sulu\Bundle\ContactBundle\Entity\AccountContact;
+use Sulu\Bundle\ContactBundle\Api\Contact as ApiContact;
 use Sulu\Bundle\ContactBundle\Entity\Contact;
 use Sulu\Bundle\ContactBundle\Entity\Fax;
 use Sulu\Bundle\ContactBundle\Entity\Email;
@@ -53,7 +55,12 @@ class ContactController extends AbstractContactController
     protected static $positionEntityName = 'SuluContactBundle:Position';
     protected static $addressEntityName = 'SuluContactBundle:Address';
     protected static $contactAddressEntityName = 'SuluContactBundle:ContactAddress';
-
+    
+    // serialization groups for contact
+    protected static $contactSerializationGroups = array(
+        'fullContact', 'partialAccount', 'partialTag', 'partialMedia', 'partialCategory'
+    );
+    
     /**
      * @var string
      */
@@ -384,6 +391,9 @@ class ContactController extends AbstractContactController
      */
     public function cgetAction(Request $request)
     {
+        $serializationGroups = array();
+        $locale = $this->getLocale($request);
+        
         if ($request->get('flat') == 'true') {
             /** @var RestHelperInterface $restHelper */
             $restHelper = $this->getRestHelper();
@@ -404,17 +414,42 @@ class ContactController extends AbstractContactController
                 $listBuilder->getLimit(),
                 $listBuilder->count()
             );
-
         } else {
             if ($request->get('bySystem') == true) {
                 $contacts = $this->getContactsByUserSystem();
+                $serializationGroups[] = 'select';
+               
             } else {
                 $contacts = $this->getDoctrine()->getRepository(self::$entityName)->findAll();
+                $serializationGroups = array_merge($serializationGroups, 
+                    static::$contactSerializationGroups
+                );
             }
-            $list = new CollectionRepresentation($contacts, self::$entityKey);
+            // convert to api-contacts
+            $apiContacts = array();
+            foreach ($contacts as $contact) {
+                $apiContacts[] = new ApiContact($contact, $locale);
+            }
+            
+            $exclusion = null;
+            if (count($serializationGroups) > 0) {
+                $exclusion = new Exclusion($serializationGroups);    
+            }
+            
+            $list = new CollectionRepresentation($apiContacts, self::$entityKey, null, $exclusion, $exclusion);
         }
+
         $view = $this->view($list, 200);
 
+        // set serialization groups
+        if (count($serializationGroups) > 0) {
+            $view->setSerializationContext(
+                SerializationContext::create()->setGroups(
+                    $serializationGroups
+                )
+            );
+        }
+        
         return $this->handleView($view);
     }
 
@@ -507,7 +542,7 @@ class ContactController extends AbstractContactController
 
             $view->setSerializationContext(
                 SerializationContext::create()->setGroups(
-                    array('fullContact', 'partialAccount', 'partialTag', 'partialMedia', 'partialCategory')
+                    static::$contactSerializationGroups
                 )
             );
         } catch (EntityNotFoundException $enfe) {
@@ -602,7 +637,7 @@ class ContactController extends AbstractContactController
             $view = $this->view($apiContact, 200);
             $view->setSerializationContext(
                 SerializationContext::create()->setGroups(
-                    array('fullContact', 'partialAccount', 'partialTag', 'partialMedia', 'partialCategory')
+                    static::$contactSerializationGroups
                 )
             );
         } catch (EntityNotFoundException $enfe) {
@@ -703,7 +738,7 @@ class ContactController extends AbstractContactController
                 $view = $this->view($apiContact, 200);
                 $view->setSerializationContext(
                     SerializationContext::create()->setGroups(
-                        array('fullContact', 'partialAccount', 'partialTag', 'partialMedia', 'partialCategory')
+                        static::$contactSerializationGroups
                     )
                 );
             }