+++ src/Sulu/Bundle/ContentBundle/Content/InternalLinksContainer.php
@@ -20,7 +20,6 @@ use Sulu\Component\Util\ArrayableInterface;
 
 /**
  * Container for InternalLinks, holds the config for a internal links, and lazy loads the structures
- * @package Sulu\Bundle\ContentBundle\Content
  */
 class InternalLinksContainer implements ArrayableInterface
 {
@@ -117,7 +116,7 @@ class InternalLinksContainer implements ArrayableInterface
             $this->contentQueryBuilder->init(
                 array(
                     'ids' => $this->ids,
-                    'properties' => (isset($this->params['properties']) ? $this->params['properties'] : array())
+                    'properties' => (isset($this->params['properties']) ? $this->params['properties']->getValue() : array())
                 )
             );
             $pages = $this->contentQueryExecutor->execute(