+++ src/Sulu/Bundle/ContentBundle/Tests/WebTests/Controller/NodeControllerTest.php
@@ -1009,7 +1009,26 @@ class NodeControllerTest extends SuluTestCase
 
         $client->request(
             'POST',
-            '/api/nodes/' . $data[3]['id'] . '?webspace=sulu_io&language=en&action=order&destination=' . $data[0]['id']
+            '/api/nodes/' . $data[1]['id'] . '?webspace=sulu_io&language=en&action=order',
+            array(
+                'position' => 3
+            )
+        );
+        $this->assertEquals(200, $client->getResponse()->getStatusCode());
+        $response = json_decode($client->getResponse()->getContent(), true);
+
+        // check some properties
+        $this->assertEquals($data[1]['id'], $response['id']);
+        $this->assertEquals('test2', $response['title']);
+        $this->assertEquals('/test2', $response['path']);
+        $this->assertEquals('/test2', $response['url']);
+
+        $client->request(
+            'POST',
+            '/api/nodes/' . $data[3]['id'] . '?webspace=sulu_io&language=en&action=order',
+            array(
+                'position' => 1
+            )
         );
         $this->assertEquals(200, $client->getResponse()->getStatusCode());
         $response = json_decode($client->getResponse()->getContent(), true);
@@ -1020,19 +1039,6 @@ class NodeControllerTest extends SuluTestCase
         $this->assertEquals('/test4', $response['path']);
         $this->assertEquals('/test4', $response['url']);
 
-        $client->request(
-            'POST',
-            '/api/nodes/' . $data[2]['id'] . '?webspace=sulu_io&language=en&action=order&destination=' . $data[3]['id']
-        );
-        $this->assertEquals(200, $client->getResponse()->getStatusCode());
-        $response = json_decode($client->getResponse()->getContent(), true);
-
-        // check some properties
-        $this->assertEquals($data[2]['id'], $response['id']);
-        $this->assertEquals('test3', $response['title']);
-        $this->assertEquals('/test3', $response['path']);
-        $this->assertEquals('/test3', $response['url']);
-
         // get child nodes from root
         $client->request('GET', '/api/nodes?depth=1&webspace=sulu_io&language=en');
         $this->assertEquals(200, $client->getResponse()->getStatusCode());
@@ -1040,9 +1046,9 @@ class NodeControllerTest extends SuluTestCase
         $items = $response['_embedded']['nodes'];
 
         $this->assertEquals(4, sizeof($items));
-        $this->assertEquals('test3', $items[0]['title']);
-        $this->assertEquals('test4', $items[1]['title']);
-        $this->assertEquals('test1', $items[2]['title']);
+        $this->assertEquals('test4', $items[0]['title']);
+        $this->assertEquals('test1', $items[1]['title']);
+        $this->assertEquals('test3', $items[2]['title']);
         $this->assertEquals('test2', $items[3]['title']);
     }
 
@@ -1061,12 +1067,15 @@ class NodeControllerTest extends SuluTestCase
 
         $client->request(
             'POST',
-            '/api/nodes/123-123-123?webspace=sulu_io&language=en&action=order&destination=' . $data[0]['id']
+            '/api/nodes/123-123-123?webspace=sulu_io&language=en&action=order',
+            array(
+                'position' => 1
+            )
         );
         $this->assertEquals(400, $client->getResponse()->getStatusCode());
     }
 
-    public function testOrderNonExistingDestination()
+    public function testOrderNonExistingPosition()
     {
         $data = array(
             array(
@@ -1081,7 +1090,10 @@ class NodeControllerTest extends SuluTestCase
 
         $client->request(
             'POST',
-            '/api/nodes/' . $data[0]['id'] . '?webspace=sulu_io&language=en&action=order&destination=123-123-123'
+            '/api/nodes/' . $data[0]['id'] . '?webspace=sulu_io&language=en&action=order',
+            array(
+                'position' => 42
+            )
         );
         $this->assertEquals(400, $client->getResponse()->getStatusCode());
     }