+++ src/Sulu/Bundle/MediaBundle/Tests/Functional/Controller/MediaControllerTest.php
@@ -13,16 +13,16 @@ namespace Sulu\Bundle\MediaBundle\Tests\Functional\Controller;
 use DateTime;
 use Doctrine\ORM\EntityManager;
 use Sulu\Bundle\MediaBundle\Entity\Collection;
-use Sulu\Bundle\MediaBundle\Entity\CollectionType;
 use Sulu\Bundle\MediaBundle\Entity\CollectionMeta;
+use Sulu\Bundle\MediaBundle\Entity\CollectionType;
+use Sulu\Bundle\MediaBundle\Entity\File;
 use Sulu\Bundle\MediaBundle\Entity\FileVersion;
 use Sulu\Bundle\MediaBundle\Entity\FileVersionMeta;
 use Sulu\Bundle\MediaBundle\Entity\Media;
 use Sulu\Bundle\MediaBundle\Entity\MediaType;
-use Sulu\Bundle\MediaBundle\Entity\File;
 use Sulu\Bundle\TagBundle\Entity\Tag;
-use Symfony\Component\HttpFoundation\File\UploadedFile;
 use Sulu\Bundle\TestBundle\Testing\SuluTestCase;
+use Symfony\Component\HttpFoundation\File\UploadedFile;
 
 class MediaControllerTest extends SuluTestCase
 {
@@ -40,6 +40,11 @@ class MediaControllerTest extends SuluTestCase
      * @var Collection
      */
     private $collection;
+
+    /**
+     * @var CollectionMeta
+     */
+    private $collectionMeta;
 
     /**
      * @var MediaType
@@ -212,13 +217,13 @@ class MediaControllerTest extends SuluTestCase
         $this->collection->setType($this->collectionType);
 
         // Collection Meta 1
-        $collectionMeta = new CollectionMeta();
-        $collectionMeta->setTitle('Test Collection');
-        $collectionMeta->setDescription('This Description is only for testing');
-        $collectionMeta->setLocale('en-gb');
-        $collectionMeta->setCollection($this->collection);
+        $this->collectionMeta = new CollectionMeta();
+        $this->collectionMeta->setTitle('Test Collection');
+        $this->collectionMeta->setDescription('This Description is only for testing');
+        $this->collectionMeta->setLocale('en-gb');
+        $this->collectionMeta->setCollection($this->collection);
 
-        $this->collection->addMeta($collectionMeta);
+        $this->collection->addMeta($this->collectionMeta);
 
         // Collection Meta 2
         $collectionMeta2 = new CollectionMeta();
@@ -231,7 +236,7 @@ class MediaControllerTest extends SuluTestCase
 
         $this->em->persist($this->collection);
         $this->em->persist($this->collectionType);
-        $this->em->persist($collectionMeta);
+        $this->em->persist($this->collectionMeta);
         $this->em->persist($collectionMeta2);
     }
 
@@ -627,7 +632,7 @@ class MediaControllerTest extends SuluTestCase
 
         $client->request(
             'POST',
-            '/api/media/' . $media->getId(),
+            '/api/media/' . $media->getId() . '?action=new-version',
             array(
                 'collection' => $this->collection->getId(),
                 'locale' => 'en-gb',
@@ -734,7 +739,7 @@ class MediaControllerTest extends SuluTestCase
 
         $client->request(
             'POST',
-            '/api/media/' . $media->getId(),
+            '/api/media/' . $media->getId() . '?action=new-version',
             array(
                 'collection' => $this->collection->getId()
             ),
@@ -858,6 +863,78 @@ class MediaControllerTest extends SuluTestCase
         $this->assertEquals($this->mediaDefaultTitle, $response->title);
         $this->assertEquals('3', $response->downloadCounter);
         $this->assertEquals($this->mediaDefaultDescription, $response->description);
+    }
+
+    /**
+     * @description Test move action
+     */
+    public function testMove()
+    {
+        $destCollection = new Collection();
+        $style = array(
+            'type' => 'circle',
+            'color' => '#ffcc00'
+        );
+
+        $destCollection->setStyle(json_encode($style));
+
+        $destCollection->setCreated(new DateTime());
+        $destCollection->setChanged(new DateTime());
+        $destCollection->setType($this->collectionType);
+        $destCollection->addMeta($this->collectionMeta);
+
+        $this->em->persist($destCollection);
+        $this->em->flush();
+
+        $media = $this->createMedia('photo');
+
+        $client = $this->createAuthenticatedClient();
+        $client->request(
+            'POST',
+            '/api/media/' . $media->getId() . '?action=move&destination=' . $destCollection->getId()
+        );
+
+        $response = json_decode($client->getResponse()->getContent(), true);
+        $this->assertEquals(200, $client->getResponse()->getStatusCode());
+        $this->assertEquals($destCollection->getId(), $response['collection']);
+        $this->assertEquals($this->mediaDefaultTitle, $response['title']);
+    }
+
+    /**
+     * @description Test move to non existing collection
+     */
+    public function testMoveNonExistingCollection()
+    {
+        $media = $this->createMedia('photo');
+
+        $client = $this->createAuthenticatedClient();
+        $client->request('POST', '/api/media/' . $media->getId() . '?action=move&destination=404');
+
+        $this->assertEquals(400, $client->getResponse()->getStatusCode());
+    }
+
+    /**
+     * @description Test move to non existing media
+     */
+    public function testMoveNonExistingMedia()
+    {
+        $client = $this->createAuthenticatedClient();
+        $client->request('POST', '/api/media/404?action=move&destination=' . $this->collection->getId());
+
+        $this->assertEquals(404, $client->getResponse()->getStatusCode());
+    }
+
+    /**
+     * @description Test non existing action
+     */
+    public function testMoveNonExistingAction()
+    {
+        $media = $this->createMedia('photo');
+
+        $client = $this->createAuthenticatedClient();
+        $client->request('POST', '/api/media/' . $media->getId() . '?action=test');
+
+        $this->assertEquals(400, $client->getResponse()->getStatusCode());
     }
 
     /**