+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 CHANGELOG for Sulu
 ==================
 
+* 0.16.0 (2015-02-24)
+    * BUGFIX      #866 [ContactBundle]   Serialization group "select" for serializing system users
+    * BUGFIX      #860 [AdminBundle]     Extended toolbar to accept more options
+    * BUGFIX      #--- [ContentBundle]   Added validation for time field
+    * BUGFIX      #865 [ContentBundle]   Added validation and localized formatted value for time field
+    * BUGFIX      #848 [ContactBundle]   Refactored delete dialog function to make it reuseable
+    * BUGFIX      #846 [MediaBundle]     Added missing dot to create event name method (\cc Daniel)
+    * ENHANCEMENT #841 [SecurityBundle]  Unique email per user
+    * BUGFIX      #837 [AdminBundle]     Javascript function for croping labels with a certain tag this.sandbox.sulu.cropAllLabels(className)
+    * ENHANCEMENT #818 [ContentBundle]   Enhanced column-navigation ordering ui
+    * BUGFIX      #857 [ContentBundle]   Added links without save could not be removed
+    * FEATURE     #789 [ContentBundle]   Added present as to smart content config (see [here ...](https://github.com/sulu-cmf/docs/blob/master/developer-documentation/300-webspaces/smart-content.md))
+    * BUGFIX      #856 [ContentBundle]   Added default values for smart content view vars
+
 * 0.15.2 (2015-02-19)
     * BUGFIX      #846 [MediaBundle]     Added missing dot to create event name method (\cc Daniel)
 
@@ -12,7 +26,6 @@ CHANGELOG for Sulu
 
 * 0.15.0 (2015-02-13)
     * BUGFIX      #833 [AdminBundle]     Added new husky version
-    * ENHANCEMENT #818 [ContentBundle]   Enhanced column-navigation ordering ui
     * BUGFIX      #829 [ContactBundle]   Account-Contacts: show full-name of contact
     * ENHANCEMENT #828 [ContactBundle]   Changed columns for contact list and made concatenated columns not sortable
     * BUGFIX      #825 [WebsiteBundle]   Fixed syntax error in ExceptionController
@@ -24,11 +37,12 @@ CHANGELOG for Sulu
     * ENHANCEMENT #735 [CategoryBundle]  Use parameters instead of FCQN of entities in service config
     * ENHANCEMENT #735 [MediaBundle]     Use parameters instead of FCQN of entities in service config
     * ENHANCEMENT #735 [TagBundle]       Use parameters instead of FCQN of entities in service config
-    * FEATURE     #792 [ContactBundle]   added widget to show all companys of contact
     * FEATURE     #820 [ContactBundle]   Contact-Import: define multiple tags: 'account_tag1 ..n'
-    * FEATURE     #810 [ContactBundle]   added command line tool for detecting missing country codes in import csv files
+    * FEATURE     #810 [ContactBundle]   Added command line tool for detecting missing country codes in import csv files
                                          that uses google geo api for finding the correct country code
-    * BUGFIX      #801 [All]             Removed unused clean task which is deleting the public directory when executed
+    * FEATURE     #805 [MediaBundle]     Implementing media move
+    * FEATURE     #792 [ContactBundle]   Added widget to show all companys of contact
+    * BUGFIX      #801 [ALL]             Removed unused clean task which is deleting the public directory when executed
                                          due to the symfony 2.6 changes to symlinks
     * FEATURE     #793 [SecurityBundle]  Added field passwordForgetToken to BaseUser-Entity
     * FEATURE     #793 [ContactBundle]   Added Repository service for Contact
@@ -37,6 +51,8 @@ CHANGELOG for Sulu
     * BUGFIX      #774 [ContentBundle]   Enabled save shadow for index pages
     * BUGFIX      #778 [ContentBundle]   Fixed shadow page with internal link and smart-content
     * BUGFIX      #790 [WebsiteBundle]   Fixed twig variables for 404 page
+    * FEATURE     #791 [ContentBundle]   Added Changelog to settings tab
+    * FEATURE     #789 [ContentBundle]   Enabled property parameters to have metadata for localization
     * FEATURE     #684 [ContentBundle]   Refactored preview to use new websocket component and only one socket for form
                                          and preview
     * FEATURE     #684 [WebsocketBundle] Implemented Websocket Component to standardize Websocket implementations